+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: dicarval <dicarval@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/26 09:53:51 by dicarval          #+#    #+#             */
-/*   Updated: 2024/04/29 16:43:42 by dicarval         ###   ########.fr       */
+/*   Updated: 2024/04/30 16:43:02 by dicarval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ int	ft_print_char(int c);
 int	ft_print_digit(long dgt);
 int	ft_print_str(char *str);
 int	ft_format_type(char type, va_list ap);
-int	ft_print_hexa(long int nbr, char type_hexa);
+int	ft_print_hexa(unsigned long long int nbr, char type_hexa);
 int	ft_print_adr(void *adr);
 
 #endif
